+++ CHANGELOG.md
@@ -10,8 +10,9 @@ Breaking Change
 - Modified program to use the mysql_class v4.0.0 version.  The v4.0.0 replaces the MySQLdb support library with the mysql.connector support library.
 
 ### Changed
+- chk_slv_time:  Replaced section of code with call to \_process_time_lag.
 - chk_slv_time:  Removed "mongo_libs.json_prt_ins_2_db" and replaced with own internal code to do the same thing.
-- chk_slv_other:  Replaced section of code with call to _chk_other.
+- chk_slv_other:  Replaced section of code with call to \_chk_other.
 - call_run_chk:  Intersect args_array & func_dict to run options.
 - main:  Refactored initial program checks.
 - run_program:  Added \*\*kwargs to argument list.
@@ -42,7 +43,8 @@ Breaking Change
 - run_program:  Fixed problem with mutable default arguments issue.
 - 
 ### Added
-- _chk_other:  Private function for chk_slv_other().  Print any possible problems found.
+- \_process_time_lag:  Private function for chk_slv_time().  Process time lag for slave.
+- \_chk_other:  Private function for chk_slv_other().  Print any possible problems found.
 
 
 ## [2.0.1] - 2018-12-06